\name{HOFS-package}
\alias{HOFS-package}
\docType{package}
\title{
  Higher Order Mutual Information Approximation for Feature Selection
}
\description{
A higher order Mutual Information (MI) based approximation technique. Instead of producing a single list of features, method produces a ranked collection of feature subsets that maximizes MI, giving better comprehension (feature ranking) as to which features work best together when selected,due to their underlying interdependent structure.
}

\author{
Krzysztof Gajowniczek

Maintainer: krzysztof_gajowniczek@sggw.edu.pl
}
\references{

}
\keyword{ package }
\seealso{
  \code{\link{HOFS}}
}
\examples{
  \dontrun{

  }
}
